refactor: update import paths for MessageDb and MessageController
